Add support for sparse datastructures & blobs
