+++ codegen-serde/src/main/kotlin/software/amazon/smithy/rust/codegen/serde/SerializeImplGenerator.kt
@@ -6,7 +6,7 @@
 package software.amazon.smithy.rust.codegen.serde
 
 import software.amazon.smithy.codegen.core.Symbol
-import software.amazon.smithy.model.neighbor.Walker
+import software.amazon.smithy.model.shapes.BlobShape
 import software.amazon.smithy.model.shapes.CollectionShape
 import software.amazon.smithy.model.shapes.DocumentShape
 import software.amazon.smithy.model.shapes.MapShape
@@ -18,13 +18,6 @@ import software.amazon.smithy.model.shapes.StructureShape
 import software.amazon.smithy.model.shapes.TimestampShape
 import software.amazon.smithy.model.shapes.UnionShape
 import software.amazon.smithy.model.traits.SensitiveTrait
-import software.amazon.smithy.rust.codegen.client.smithy.ClientCodegenContext
-import software.amazon.smithy.rust.codegen.client.smithy.customize.ClientCodegenDecorator
-import software.amazon.smithy.rust.codegen.core.rustlang.Attribute
-import software.amazon.smithy.rust.codegen.core.rustlang.CargoDependency
-import software.amazon.smithy.rust.codegen.core.rustlang.DependencyScope
-import software.amazon.smithy.rust.codegen.core.rustlang.Feature
-import software.amazon.smithy.rust.codegen.core.rustlang.RustModule
 import software.amazon.smithy.rust.codegen.core.rustlang.RustWriter
 import software.amazon.smithy.rust.codegen.core.rustlang.Writable
 import software.amazon.smithy.rust.codegen.core.rustlang.map
@@ -36,58 +29,18 @@ import software.amazon.smithy.rust.codegen.core.rustlang.withBlock
 import software.amazon.smithy.rust.codegen.core.rustlang.writable
 import software.amazon.smithy.rust.codegen.core.smithy.CodegenContext
 import software.amazon.smithy.rust.codegen.core.smithy.RuntimeType
-import software.amazon.smithy.rust.codegen.core.smithy.RuntimeType.Companion.preludeScope
-import software.amazon.smithy.rust.codegen.core.smithy.RustCrate
 import software.amazon.smithy.rust.codegen.core.smithy.contextName
 import software.amazon.smithy.rust.codegen.core.smithy.generators.UnionGenerator
 import software.amazon.smithy.rust.codegen.core.smithy.generators.renderUnknownVariant
 import software.amazon.smithy.rust.codegen.core.smithy.isOptional
 import software.amazon.smithy.rust.codegen.core.smithy.protocols.shapeFunctionName
 import software.amazon.smithy.rust.codegen.core.smithy.rustType
+import software.amazon.smithy.rust.codegen.core.util.PANIC
 import software.amazon.smithy.rust.codegen.core.util.dq
 import software.amazon.smithy.rust.codegen.core.util.hasTrait
 import software.amazon.smithy.rust.codegen.core.util.isTargetUnit
 import software.amazon.smithy.rust.codegen.core.util.letIf
 import software.amazon.smithy.rust.codegen.core.util.toPascalCase
-
-val SerdeFeature = Feature("serde", false, listOf("dep:serde"))
-val Module =
-    RustModule.public(
-        "serde_impl",
-        additionalAttributes = listOf(Attribute.featureGate(SerdeFeature.name)),
-        documentationOverride = "Implementations of `serde` for model types",
-    )
-
-class KotlinClientSerdeDecorator : ClientCodegenDecorator {
-    override val name: String = "ClientSerdeDecorator"
-    override val order: Byte = 0
-
-    override fun extras(
-        codegenContext: ClientCodegenContext,
-        rustCrate: RustCrate,
-    ) {
-        rustCrate.mergeFeature(SerdeFeature)
-        val generator = SerializeImplGenerator(codegenContext)
-        rustCrate.withModule(Module) {
-            serializationRoots(codegenContext).forEach {
-                generator.generateRootSerializerForShape(
-                    it,
-                )(this)
-            }
-            addDependency(SupportStructures.serializeRedacted().toSymbol())
-            addDependency(SupportStructures.serializeUnredacted().toSymbol())
-        }
-    }
-}
-
-/**
- * All entry points for serialization in the service closure.
- */
-fun serializationRoots(ctx: CodegenContext): List<Shape> {
-    val serviceShape = ctx.serviceShape
-    val walker = Walker(ctx.model)
-    return walker.walkShapes(serviceShape).filter { it.hasTrait<SerdeTrait>() }
-}
 
 class SerializeImplGenerator(private val codegenContext: CodegenContext) {
     fun generateRootSerializerForShape(shape: Shape): Writable = serializerFn(shape, null)
@@ -111,6 +64,7 @@ class SerializeImplGenerator(private val codegenContext: CodegenContext) {
                 is StructureShape -> RuntimeType.forInlineFun(name, Module, structSerdeImpl(shape))
                 is UnionShape -> RuntimeType.forInlineFun(name, Module, serializeUnionImpl(shape))
                 is TimestampShape -> serializeDateTime(shape)
+                is BlobShape -> serializeBlob(shape)
                 is StringShape, is NumberShape -> directSerde(shape)
                 is DocumentShape -> serializeDocument(shape)
                 else -> null
@@ -122,7 +76,7 @@ class SerializeImplGenerator(private val codegenContext: CodegenContext) {
                     shape is MapShape -> serializeMap(shape)
                     shape is CollectionShape -> serializeList(shape)
                     // Need to figure out the best default here.
-                    else -> serializeWithTodo(shape)
+                    else -> PANIC("No serializer supported for $shape")
                 }
             if (wrapper != null && applyTo != null) {
                 rustTemplate("&#{wrapper}(#{applyTo})", "wrapper" to wrapper, "applyTo" to applyTo)
@@ -132,14 +86,6 @@ class SerializeImplGenerator(private val codegenContext: CodegenContext) {
             }
         }
     }
-
-    private fun serializeWithTodo(shape: Shape): RuntimeType =
-        serializeWithWrapper(shape) { _ ->
-            // PANIC("cant serialize $shape")
-            writable {
-                rust("serializer.serialize_str(\"todo\")")
-            }
-        }
 
     private fun serializeMap(shape: MapShape): RuntimeType =
         serializeWithWrapper(shape) { value ->
@@ -299,15 +245,9 @@ class SerializeImplGenerator(private val codegenContext: CodegenContext) {
                             )
                         }
                     if (codegenContext.symbolProvider.toSymbol(member).isOptional()) {
-                        rustTemplate(
-                            "if let Some($field) = &inner.$fieldName { #{serializeField} }",
-                            "serializeField" to fieldSerialization,
-                        )
+                        rust("if let Some($field) = &inner.$fieldName { #T }", fieldSerialization)
                     } else {
-                        rustTemplate(
-                            "let $field = &inner.$fieldName; #{serializeField}",
-                            "serializeField" to fieldSerialization,
-                        )
+                        rust("let $field = &inner.$fieldName; #T", fieldSerialization)
                     }
                 }
                 rust("s.end()")
@@ -359,6 +299,22 @@ class SerializeImplGenerator(private val codegenContext: CodegenContext) {
         RuntimeType.forInlineFun("SerializeDateTime", Module) {
             implSerializeConfigured(codegenContext.symbolProvider.toSymbol(shape)) {
                 rust("serializer.serialize_str(&self.value.to_string())")
+            }
+        }
+
+    private fun serializeBlob(shape: BlobShape): RuntimeType =
+        RuntimeType.forInlineFun("SerializeBlob", Module) {
+            implSerializeConfigured(codegenContext.symbolProvider.toSymbol(shape)) {
+                rustTemplate(
+                    """
+                    if serializer.is_human_readable() {
+                        serializer.serialize_str(&#{base64_encode}(&self.value.as_ref()))
+                    } else {
+                        serializer.serialize_bytes(&self.value.as_ref())
+                    }
+                    """,
+                    "base64_encode" to RuntimeType.base64Encode(codegenContext.runtimeConfig),
+                )
             }
         }
 
@@ -443,212 +399,4 @@ class SerializeImplGenerator(private val codegenContext: CodegenContext) {
             )
         }
     }
-}
-
-object SupportStructures {
-    private val supportModule =
-        RustModule.public("support", Module, documentationOverride = "Support traits and structures for serde")
-
-    private val serde = CargoDependency.Serde.copy(scope = DependencyScope.Compile, optional = true).toType()
-
-    val codegenScope =
-        arrayOf(
-            *preludeScope,
-            "ConfigurableSerde" to configurableSerde(),
-            "SerializeConfigured" to serializeConfigured(),
-            "ConfigurableSerdeRef" to configurableSerdeRef(),
-            "SerializationSettings" to serializationSettings(),
-            "Sensitive" to sensitive(),
-            "serde" to serde,
-            "serialize_redacted" to serializeRedacted(),
-            "serialize_unredacted" to serializeUnredacted(),
-        )
-
-    fun serializeRedacted(): RuntimeType =
-        RuntimeType.forInlineFun("serialize_redacted", supportModule) {
-            rustTemplate(
-                """
-                /// Serialize a value redacting sensitive fields
-                ///
-                /// This function is intended to be used by `serde(serialize_with = "serialize_redacted")`
-                pub fn serialize_redacted<'a, T, S: #{serde}::Serializer>(value: &'a T, serializer: S) -> Result<S::Ok, S::Error>
-                where
-                    T: #{SerializeConfigured},
-                {
-                    use #{serde}::Serialize;
-                    value
-                        .serialize_ref(&#{SerializationSettings} { redact_sensitive_fields: true })
-                        .serialize(serializer)
-                }
-                """,
-                "serde" to serde,
-                "SerializeConfigured" to serializeConfigured(),
-                "SerializationSettings" to serializationSettings(),
-            )
-        }
-
-    fun serializeUnredacted(): RuntimeType =
-        RuntimeType.forInlineFun("serialize_unredacted", supportModule) {
-            rustTemplate(
-                """
-                /// Serialize a value without redacting sensitive fields
-                ///
-                /// This function is intended to be used by `serde(serialize_with = "serialize_unredacted")`
-                pub fn serialize_unredacted<'a, T, S: #{serde}::Serializer>(value: &'a T, serializer: S) -> Result<S::Ok, S::Error>
-                where
-                    T: #{SerializeConfigured},
-                {
-                    use #{serde}::Serialize;
-                    value
-                        .serialize_ref(&#{SerializationSettings} { redact_sensitive_fields: false })
-                        .serialize(serializer)
-                }
-                """,
-                "serde" to serde,
-                "SerializeConfigured" to serializeConfigured(),
-                "SerializationSettings" to serializationSettings(),
-            )
-        }
-
-    private fun serializeConfigured(): RuntimeType =
-        RuntimeType.forInlineFun("SerializeConfigured", supportModule) {
-            rustTemplate(
-                """
-                /// Trait that allows configuring serialization
-                /// **This trait should not be implemented directly!** Instead, `impl Serialize for ConfigurableSerdeRef<T>`
-                pub trait SerializeConfigured {
-                    /// Return a `Serialize` implementation for this object that owns the object.
-                    ///
-                    /// Use this if you need to create `Arc<dyn Serialize>` or similar.
-                    fn serialize_owned(self, settings: #{SerializationSettings}) -> impl #{serde}::Serialize;
-
-                    /// Return a `Serialize` implementation for this object that borrows from the given object
-                    fn serialize_ref<'a>(&'a self, settings: &'a #{SerializationSettings}) -> impl #{serde}::Serialize + 'a;
-                }
-
-                /// Blanket implementation for all `T` that implement `ConfigurableSerdeRef`
-                impl<T> SerializeConfigured for T
-                where
-                    for<'a> #{ConfigurableSerdeRef}<'a, T>: #{serde}::Serialize,
-                {
-                    fn serialize_owned(
-                        self,
-                        settings: #{SerializationSettings},
-                    ) -> impl #{serde}::Serialize {
-                        #{ConfigurableSerde} {
-                            value: self,
-                            settings,
-                        }
-                    }
-
-                    fn serialize_ref<'a>(
-                        &'a self,
-                        settings: &'a #{SerializationSettings},
-                    ) -> impl #{serde}::Serialize + 'a {
-                        #{ConfigurableSerdeRef} { value: self, settings }
-                    }
-                }
-                """,
-                "ConfigurableSerde" to configurableSerde(),
-                "ConfigurableSerdeRef" to configurableSerdeRef(),
-                "SerializationSettings" to serializationSettings(),
-                "serde" to serde,
-            )
-        }
-
-    private fun sensitive() =
-        RuntimeType.forInlineFun("Sensitive", supportModule) {
-            rustTemplate(
-                """
-                pub(crate) struct Sensitive<T>(pub(crate) T);
-
-                impl<'a, T> ::serde::Serialize for ConfigurableSerdeRef<'a, Sensitive<T>>
-                where
-                T: #{serde}::Serialize,
-                {
-                    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-                    where
-                    S: serde::Serializer,
-                    {
-                        match self.settings.redact_sensitive_fields {
-                            true => serializer.serialize_str("<redacted>"),
-                            false => self.value.0.serialize(serializer),
-                        }
-                    }
-                }
-                """,
-                "serde" to CargoDependency.Serde.toType(),
-            )
-        }
-
-    private fun configurableSerde() =
-        RuntimeType.forInlineFun("ConfigurableSerde", supportModule) {
-            rustTemplate(
-                """
-                ##[allow(missing_docs)]
-                pub(crate) struct ConfigurableSerde<T> {
-                    pub(crate) value: T,
-                    pub(crate) settings: #{SerializationSettings}
-                }
-
-                impl<T> #{serde}::Serialize for ConfigurableSerde<T> where for <'a> ConfigurableSerdeRef<'a, T>: #{serde}::Serialize {
-                    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-                    where
-                        S: #{serde}::Serializer,
-                    {
-                        #{ConfigurableSerdeRef} {
-                            value: &self.value,
-                            settings: &self.settings,
-                        }
-                        .serialize(serializer)
-                    }
-                }
-
-                """,
-                "SerializationSettings" to serializationSettings(),
-                "ConfigurableSerdeRef" to configurableSerdeRef(),
-                "serde" to CargoDependency.Serde.toType(),
-            )
-        }
-
-    private fun configurableSerdeRef() =
-        RuntimeType.forInlineFun("ConfigurableSerdeRef", supportModule) {
-            rustTemplate(
-                """
-                ##[allow(missing_docs)]
-                pub(crate) struct ConfigurableSerdeRef<'a, T> {
-                    pub(crate) value: &'a T,
-                    pub(crate) settings: &'a SerializationSettings
-                }
-                """,
-            )
-        }
-
-    private fun serializationSettings() =
-        RuntimeType.forInlineFun("SerializationSettings", supportModule) {
-            // TODO(serde): Add a builder for this structure and make it non-exhaustive
-            // TODO(serde): Consider removing `derive(Default)`
-            rustTemplate(
-                """
-                /// Settings for use when serializing structures
-                ##[non_exhaustive]
-                ##[derive(Copy, Clone, Debug, Default)]
-                pub struct SerializationSettings {
-                    /// Replace all sensitive fields with `<redacted>` during serialization
-                    pub redact_sensitive_fields: bool,
-                }
-
-                impl SerializationSettings {
-                    /// Replace all `@sensitive` fields with `<redacted>` when serializing.
-                    ///
-                    /// Note: This may alter the type of the serialized output and make it impossible to deserialize as
-                    /// numerical fields will be replaced with strings.
-                    pub fn redact_sensitive_fields() -> Self { Self { redact_sensitive_fields: true } }
-
-                    /// Preserve the contents of sensitive fields during serializing
-                    pub fn leak_sensitive_fields() -> Self { Self { redact_sensitive_fields: false } }
-                }
-                """,
-            )
-        }
 }
